+++ compiler/lib-dynlink/stubs.c
@@ -1,17 +1,17 @@
 #include <caml/misc.h>
 
-void jsoo_get_bytecode_sections () {
+CAMLweakdef void jsoo_get_bytecode_sections () {
   caml_fatal_error("Unimplemented Javascript primitive jsoo_get_bytecode_sections!");
 }
 
-void jsoo_get_runtime_aliases () {
+CAMLweakdef void jsoo_get_runtime_aliases () {
   caml_fatal_error("Unimplemented Javascript primitive jsoo_get_runtime_aliases!");
 }
 
-void jsoo_toplevel_init_compile () {
+CAMLweakdef void jsoo_toplevel_init_compile () {
   caml_fatal_error("Unimplemented Javascript primitive jsoo_toplevel_init_compile!");
 }
 
-void jsoo_toplevel_init_reloc () {
+CAMLweakdef void jsoo_toplevel_init_reloc () {
   caml_fatal_error("Unimplemented Javascript primitive jsoo_toplevel_init_reloc!");
 }