+++ lib/js_of_ocaml/js_of_ocaml_stubs.c
@@ -1,17 +1,17 @@
 #include <caml/misc.h>
 
-void caml_js_error_of_exception () {
+CAMLweakdef void caml_js_error_of_exception () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_error_of_exception!");
 }
 
-void caml_js_get_console () {
+CAMLweakdef void caml_js_get_console () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_get_console!");
 }
 
-void caml_js_html_entities () {
+CAMLweakdef void caml_js_html_entities () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_html_entities!");
 }
 
-void caml_xmlhttprequest_create () {
+CAMLweakdef void caml_xmlhttprequest_create () {
   caml_fatal_error("Unimplemented Javascript primitive caml_xmlhttprequest_create!");
 }