+++ lib/lwt/graphics/graphics_js_stubs.c
@@ -1,17 +1,17 @@
 #include <caml/misc.h>
 
-void caml_gr_doc_of_state () {
+CAMLweakdef void caml_gr_doc_of_state () {
   caml_fatal_error("Unimplemented Javascript primitive caml_gr_doc_of_state!");
 }
 
-void caml_gr_state_create () {
+CAMLweakdef void caml_gr_state_create () {
   caml_fatal_error("Unimplemented Javascript primitive caml_gr_state_create!");
 }
 
-void caml_gr_state_get () {
+CAMLweakdef void caml_gr_state_get () {
   caml_fatal_error("Unimplemented Javascript primitive caml_gr_state_get!");
 }
 
-void caml_gr_state_set () {
+CAMLweakdef void caml_gr_state_set () {
   caml_fatal_error("Unimplemented Javascript primitive caml_gr_state_set!");
 }