+++ lib/runtime/js_of_ocaml_runtime_stubs.c
@@ -1,313 +1,313 @@
 #include <caml/misc.h>
 
-void bigstring_of_array_buffer () {
+CAMLweakdef void bigstring_of_array_buffer () {
   caml_fatal_error("Unimplemented Javascript primitive bigstring_of_array_buffer!");
 }
 
-void bigstring_of_typed_array () {
+CAMLweakdef void bigstring_of_typed_array () {
   caml_fatal_error("Unimplemented Javascript primitive bigstring_of_typed_array!");
 }
 
-void bigstring_to_array_buffer () {
+CAMLweakdef void bigstring_to_array_buffer () {
   caml_fatal_error("Unimplemented Javascript primitive bigstring_to_array_buffer!");
 }
 
-void bigstring_to_typed_array () {
+CAMLweakdef void bigstring_to_typed_array () {
   caml_fatal_error("Unimplemented Javascript primitive bigstring_to_typed_array!");
 }
 
-void caml_assume_no_perform () {
+CAMLweakdef void caml_assume_no_perform () {
   caml_fatal_error("Unimplemented Javascript primitive caml_assume_no_perform!");
 }
 
-void caml_ba_from_typed_array () {
+CAMLweakdef void caml_ba_from_typed_array () {
   caml_fatal_error("Unimplemented Javascript primitive caml_ba_from_typed_array!");
 }
 
-void caml_ba_kind_of_typed_array () {
+CAMLweakdef void caml_ba_kind_of_typed_array () {
   caml_fatal_error("Unimplemented Javascript primitive caml_ba_kind_of_typed_array!");
 }
 
-void caml_ba_to_typed_array () {
+CAMLweakdef void caml_ba_to_typed_array () {
   caml_fatal_error("Unimplemented Javascript primitive caml_ba_to_typed_array!");
 }
 
-void caml_bytes_of_uint8_array () {
+CAMLweakdef void caml_bytes_of_uint8_array () {
   caml_fatal_error("Unimplemented Javascript primitive caml_bytes_of_uint8_array!");
 }
 
-void caml_create_file () {
+CAMLweakdef void caml_create_file () {
   caml_fatal_error("Unimplemented Javascript primitive caml_create_file!");
 }
 
-void caml_custom_identifier () {
+CAMLweakdef void caml_custom_identifier () {
   caml_fatal_error("Unimplemented Javascript primitive caml_custom_identifier!");
 }
 
-void caml_exn_with_js_backtrace () {
+CAMLweakdef void caml_exn_with_js_backtrace () {
   caml_fatal_error("Unimplemented Javascript primitive caml_exn_with_js_backtrace!");
 }
 
-void caml_int64_create_lo_mi_hi () {
+CAMLweakdef void caml_int64_create_lo_mi_hi () {
   caml_fatal_error("Unimplemented Javascript primitive caml_int64_create_lo_mi_hi!");
 }
 
-void caml_js_call () {
+CAMLweakdef void caml_js_call () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_call!");
 }
 
-void caml_js_delete () {
+CAMLweakdef void caml_js_delete () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_delete!");
 }
 
-void caml_js_equals () {
+CAMLweakdef void caml_js_equals () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_equals!");
 }
 
-void caml_js_error_option_of_exception () {
+CAMLweakdef void caml_js_error_option_of_exception () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_error_option_of_exception!");
 }
 
-void caml_js_eval_string () {
+CAMLweakdef void caml_js_eval_string () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_eval_string!");
 }
 
-void caml_js_expr () {
+CAMLweakdef void caml_js_expr () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_expr!");
 }
 
-void caml_js_from_array () {
+CAMLweakdef void caml_js_from_array () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_from_array!");
 }
 
-void caml_js_from_bool () {
+CAMLweakdef void caml_js_from_bool () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_from_bool!");
 }
 
-void caml_js_from_float () {
+CAMLweakdef void caml_js_from_float () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_from_float!");
 }
 
-void caml_js_from_int32 () {
+CAMLweakdef void caml_js_from_int32 () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_from_int32!");
 }
 
-void caml_js_from_nativeint () {
+CAMLweakdef void caml_js_from_nativeint () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_from_nativeint!");
 }
 
-void caml_js_from_string () {
+CAMLweakdef void caml_js_from_string () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_from_string!");
 }
 
-void caml_js_fun_call () {
+CAMLweakdef void caml_js_fun_call () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_fun_call!");
 }
 
-void caml_js_get () {
+CAMLweakdef void caml_js_get () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_get!");
 }
 
-void caml_js_instanceof () {
+CAMLweakdef void caml_js_instanceof () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_instanceof!");
 }
 
-void caml_js_meth_call () {
+CAMLweakdef void caml_js_meth_call () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_meth_call!");
 }
 
-void caml_js_new () {
+CAMLweakdef void caml_js_new () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_new!");
 }
 
-void caml_js_object () {
+CAMLweakdef void caml_js_object () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_object!");
 }
 
-void caml_js_pure_expr () {
+CAMLweakdef void caml_js_pure_expr () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_pure_expr!");
 }
 
-void caml_js_set () {
+CAMLweakdef void caml_js_set () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_set!");
 }
 
-void caml_js_strict_equals () {
+CAMLweakdef void caml_js_strict_equals () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_strict_equals!");
 }
 
-void caml_js_to_array () {
+CAMLweakdef void caml_js_to_array () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_to_array!");
 }
 
-void caml_js_to_bool () {
+CAMLweakdef void caml_js_to_bool () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_to_bool!");
 }
 
-void caml_js_to_byte_string () {
+CAMLweakdef void caml_js_to_byte_string () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_to_byte_string!");
 }
 
-void caml_js_to_float () {
+CAMLweakdef void caml_js_to_float () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_to_float!");
 }
 
-void caml_js_to_int32 () {
+CAMLweakdef void caml_js_to_int32 () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_to_int32!");
 }
 
-void caml_js_to_nativeint () {
+CAMLweakdef void caml_js_to_nativeint () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_to_nativeint!");
 }
 
-void caml_js_to_string () {
+CAMLweakdef void caml_js_to_string () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_to_string!");
 }
 
-void caml_js_typeof () {
+CAMLweakdef void caml_js_typeof () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_typeof!");
 }
 
-void caml_js_var () {
+CAMLweakdef void caml_js_var () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_var!");
 }
 
-void caml_js_wrap_callback () {
+CAMLweakdef void caml_js_wrap_callback () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_wrap_callback!");
 }
 
-void caml_js_wrap_callback_arguments () {
+CAMLweakdef void caml_js_wrap_callback_arguments () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_wrap_callback_arguments!");
 }
 
-void caml_js_wrap_callback_strict () {
+CAMLweakdef void caml_js_wrap_callback_strict () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_wrap_callback_strict!");
 }
 
-void caml_js_wrap_callback_unsafe () {
+CAMLweakdef void caml_js_wrap_callback_unsafe () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_wrap_callback_unsafe!");
 }
 
-void caml_js_wrap_meth_callback () {
+CAMLweakdef void caml_js_wrap_meth_callback () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_wrap_meth_callback!");
 }
 
-void caml_js_wrap_meth_callback_arguments () {
+CAMLweakdef void caml_js_wrap_meth_callback_arguments () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_wrap_meth_callback_arguments!");
 }
 
-void caml_js_wrap_meth_callback_strict () {
+CAMLweakdef void caml_js_wrap_meth_callback_strict () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_wrap_meth_callback_strict!");
 }
 
-void caml_js_wrap_meth_callback_unsafe () {
+CAMLweakdef void caml_js_wrap_meth_callback_unsafe () {
   caml_fatal_error("Unimplemented Javascript primitive caml_js_wrap_meth_callback_unsafe!");
 }
 
-void caml_jsbytes_of_string () {
+CAMLweakdef void caml_jsbytes_of_string () {
   caml_fatal_error("Unimplemented Javascript primitive caml_jsbytes_of_string!");
 }
 
-void caml_jsoo_build_config () {
+CAMLweakdef void caml_jsoo_build_config () {
   caml_fatal_error("Unimplemented Javascript primitive caml_jsoo_build_config!");
 }
 
-void caml_jsoo_flags_effects () {
+CAMLweakdef void caml_jsoo_flags_effects () {
   caml_fatal_error("Unimplemented Javascript primitive caml_jsoo_flags_effects!");
 }
 
-void caml_jsoo_flags_use_js_string () {
+CAMLweakdef void caml_jsoo_flags_use_js_string () {
   caml_fatal_error("Unimplemented Javascript primitive caml_jsoo_flags_use_js_string!");
 }
 
-void caml_jsoo_promise_unwrap () {
+CAMLweakdef void caml_jsoo_promise_unwrap () {
   caml_fatal_error("Unimplemented Javascript primitive caml_jsoo_promise_unwrap!");
 }
 
-void caml_jsoo_promise_wrap () {
+CAMLweakdef void caml_jsoo_promise_wrap () {
   caml_fatal_error("Unimplemented Javascript primitive caml_jsoo_promise_wrap!");
 }
 
-void caml_jsoo_runtime_value () {
+CAMLweakdef void caml_jsoo_runtime_value () {
   caml_fatal_error("Unimplemented Javascript primitive caml_jsoo_runtime_value!");
 }
 
-void caml_jsstring_of_string () {
+CAMLweakdef void caml_jsstring_of_string () {
   caml_fatal_error("Unimplemented Javascript primitive caml_jsstring_of_string!");
 }
 
-void caml_list_mount_point () {
+CAMLweakdef void caml_list_mount_point () {
   caml_fatal_error("Unimplemented Javascript primitive caml_list_mount_point!");
 }
 
-void caml_list_of_js_array () {
+CAMLweakdef void caml_list_of_js_array () {
   caml_fatal_error("Unimplemented Javascript primitive caml_list_of_js_array!");
 }
 
-void caml_list_to_js_array () {
+CAMLweakdef void caml_list_to_js_array () {
   caml_fatal_error("Unimplemented Javascript primitive caml_list_to_js_array!");
 }
 
-void caml_ml_channel_redirect () {
+CAMLweakdef void caml_ml_channel_redirect () {
   caml_fatal_error("Unimplemented Javascript primitive caml_ml_channel_redirect!");
 }
 
-void caml_ml_channel_restore () {
+CAMLweakdef void caml_ml_channel_restore () {
   caml_fatal_error("Unimplemented Javascript primitive caml_ml_channel_restore!");
 }
 
-void caml_ml_set_channel_output () {
+CAMLweakdef void caml_ml_set_channel_output () {
   caml_fatal_error("Unimplemented Javascript primitive caml_ml_set_channel_output!");
 }
 
-void caml_ml_set_channel_refill () {
+CAMLweakdef void caml_ml_set_channel_refill () {
   caml_fatal_error("Unimplemented Javascript primitive caml_ml_set_channel_refill!");
 }
 
-void caml_mount_autoload () {
+CAMLweakdef void caml_mount_autoload () {
   caml_fatal_error("Unimplemented Javascript primitive caml_mount_autoload!");
 }
 
-void caml_ojs_new_arr () {
+CAMLweakdef void caml_ojs_new_arr () {
   caml_fatal_error("Unimplemented Javascript primitive caml_ojs_new_arr!");
 }
 
-void caml_pure_js_expr () {
+CAMLweakdef void caml_pure_js_expr () {
   caml_fatal_error("Unimplemented Javascript primitive caml_pure_js_expr!");
 }
 
-void caml_read_file_content () {
+CAMLweakdef void caml_read_file_content () {
   caml_fatal_error("Unimplemented Javascript primitive caml_read_file_content!");
 }
 
-void caml_string_of_array () {
+CAMLweakdef void caml_string_of_array () {
   caml_fatal_error("Unimplemented Javascript primitive caml_string_of_array!");
 }
 
-void caml_string_of_jsbytes () {
+CAMLweakdef void caml_string_of_jsbytes () {
   caml_fatal_error("Unimplemented Javascript primitive caml_string_of_jsbytes!");
 }
 
-void caml_string_of_jsstring () {
+CAMLweakdef void caml_string_of_jsstring () {
   caml_fatal_error("Unimplemented Javascript primitive caml_string_of_jsstring!");
 }
 
-void caml_string_of_uint8_array () {
+CAMLweakdef void caml_string_of_uint8_array () {
   caml_fatal_error("Unimplemented Javascript primitive caml_string_of_uint8_array!");
 }
 
-void caml_throw_js_exception () {
+CAMLweakdef void caml_throw_js_exception () {
   caml_fatal_error("Unimplemented Javascript primitive caml_throw_js_exception!");
 }
 
-void caml_uint8_array_of_bytes () {
+CAMLweakdef void caml_uint8_array_of_bytes () {
   caml_fatal_error("Unimplemented Javascript primitive caml_uint8_array_of_bytes!");
 }
 
-void caml_unmount () {
+CAMLweakdef void caml_unmount () {
   caml_fatal_error("Unimplemented Javascript primitive caml_unmount!");
 }
 
-void debugger () {
+CAMLweakdef void debugger () {
   caml_fatal_error("Unimplemented Javascript primitive debugger!");
 }